Use u8 for input/output state
This will allow us to change the internal representation of the grid simulation without breaking the other binaries.

+++ benches/benchmark.rs
@@ -6,9 +6,9 @@ const GRID_SIZE: u32 = 1024;
 fn benchmark(c: &mut Criterion) {
     // Create a random initialize state for the simulation.
     let num_cells = (GRID_SIZE * GRID_SIZE) as usize;
-    let mut init_state = vec![0u32; num_cells as usize];
+    let mut init_state = vec![0; num_cells as usize];
     for i in 0..num_cells {
-        init_state[i] = rand::random::<u32>() % 2;
+        init_state[i] = rand::random::<u8>() % 2;
     }
 
     let mut sim = pollster::block_on(LifeSimulation::new(GRID_SIZE, &init_state));
